Tab system - Merge to Main
All changes in the tab system are complete. Now I will have to work on the xyz trajectories for animations. Will merge these changes to the main branch.
